fix: width of burger layers in the constructor
When there is no scroll in the filling section,
there is an 8 pixel difference between the widths
of the bun and filling layers. We can fix this,
for example, by tracking (useRef) the list element,
when changing its width (i.e. when a scroll
appears) change the value of the right padding of
the bun layers to +8. But I didn't do it: I doubt
that this is a good solution.